Migrate more methods into traits
