+++ src/Views/Columns/Traits/HasLabelFormat.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Views\Columns\Traits;
+
+use Rappasoft\LaravelLivewireTables\Views\Column;
+
+trait HasLabelFormat
+{
+    protected mixed $labelCallback = null;
+
+    protected mixed $formatCallback = null;
+
+    public function hasFormatter(): bool
+    {
+        return $this->formatCallback !== null;
+    }
+
+    public function getFormatCallback(): ?callable
+    {
+        return $this->formatCallback;
+    }
+
+    // TODO
+    public function getLabelCallback(): ?callable
+    {
+        return $this->labelCallback;
+    }
+    
+    public function label(callable $callback): self
+    {
+        $this->from = null;
+        $this->field = null;
+        $this->labelCallback = $callback;
+
+        return $this;
+    }
+
+    public function format(callable $callable): Column
+    {
+        $this->formatCallback = $callable;
+
+        return $this;
+    }
+}
